Refactor DB test
