--- backend/internal/models/.env
@@ -1,11 +1,0 @@
-# assign each field and rename file to `.env`
-
-AVHBS_LOG_LEVEL=DEBUG # DEBUG, INFO, WARN, ERROR
-AVHBS_HTTP_PORT=8081
-AVHBS_FRONTEND_PATH=/home/maseiler/Coding/av-huette/avh-booking-system/frontend/dist
-
-DB_HOST=localhost
-DB_PORT=5432
-DB_NAME=avhbs_test
-DB_USER=postgres
-DB_PASSWORD=h5iasHm6hGF7m6J7VCy31ndDRVtWvvLYCpQwD1hG
