+++ backend/test/database/.env
@@ -1,0 +1,6 @@
+# change values according to your setup
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=avhbs_test
+DB_USER=postgres
+DB_PASSWORD=h5iasHm6hGF7m6J7VCy31ndDRVtWvvLYCpQwD1hG
